Protocol ref dev

See merge request s1810979/Netw2!14

+++ src/client/client/client.h
@@ -9,9 +9,9 @@ typedef struct {
     unsigned fd;
     struct sockaddr* sock;
     player_t* player;
-    unsigned quality;
+    uint8_t quality;
 
-    unsigned batch_nr;
+    uint32_t batch_nr;
     bool EOS_received;
 } client_t;
